requirements.txt
streamlit
ultralytics
pillow
streamlit-webrtc
av
pyttsx3
